MyPosts refactored to TS.
